+++ R/get_b0.R
@@ -37,46 +37,64 @@
 #'   response is better (used in power calculation for the overall
 #'   test only)
 #' @param sigma_error Residual variance assumed for type = "continuous"
+#' @param lambda0 Intercept of exponential regression (on non-log scale)
+#' @param cens_time Function to generate the censoring time, only
+#'   needed for data_type = "survival"
+#' @param t_mile Time point for comparing survival probabilities for overall
+#'   test for treatment effect
 #' @param interval Interval to search for b0 handed over to uniroot
 #' @return Vector of model parameter b0
 #' @export
 #' @examples
-#' 
+#'
 #' scal <- 100
-#' X <- generate_X_dist(n=500*scal, p=30, rho=0.5)
-#' trt <- generate_trt(n=500*scal, p_trt = 0.5)
+#' X <- generate_X_dist(n = 500 * scal, p = 30, rho = 0.5)
+#' trt <- generate_trt(n = 500 * scal, p_trt = 0.5)
 #' prog <- "0.5*((X1=='Y')+X11)"
 #' pred <- "X11>0.5"
-#' get_b0(X, scal, prog, pred, trt, b1 = 0.5, type="continuous",
-#'        power = 0.9, alpha = 0.025, interval=c(-2,2), sigma_error=1)
-#' get_b0(X, scal, prog, pred, trt, b1 = 0.5, type = "binary",
-#'        power = 0.9, alpha = 0.025, interval=c(-2,2))
-get_b0 <- function(X, scal, prog, pred, trt, b1, type, 
-                   power = 0.9, alpha = 0.025, 
-                   sign_better = 1, sigma_error, interval = c(-5,5)){
-  
+#' get_b0(X, scal, prog, pred, trt,
+#'   b1 = 0.5, type = "continuous",
+#'   power = 0.9, alpha = 0.025, interval = c(-2, 2), sigma_error = 1
+#' )
+#' get_b0(X, scal, prog, pred, trt,
+#'   b1 = 0.5, type = "binary",
+#'   power = 0.9, alpha = 0.025, interval = c(-2, 2)
+#' )
+get_b0 <- function(X, scal, prog, pred, trt, b1, type,
+                   power = 0.9, alpha = 0.025,
+                   sign_better = 1, sigma_error,
+                   lambda0, cens_time, t_mile, interval = c(-5, 5)) {
   stopifnot(nrow(X) == length(trt), is.character(prog), is.character(pred))
   stopifnot(type %in% c("continuous", "binary", "count", "survival"))
-  if(type %in% c("count", "survival"))
+  if (type %in% c("count")) {
     stop(sprintf("%s not implemented yet", type))
-  if(power < alpha)
+  }
+  if (power < alpha) {
     stop("Power needs to be > alpha")
-  
+  }
+
   ## generate prognostic and predictive covariates
   pred_vals <- with(X, eval(parse(text = pred)))
   prog_vals <- with(X, eval(parse(text = prog)))
-  
-  opt <- uniroot(function(b0, ...){
-    calc_power(c(b0, b1), ...)[1] - power
-  }, scal = scal, prog_vals=prog_vals, trt=trt, pred_vals=pred_vals, alpha = alpha, 
-  type = type, sign_better = sign_better, sigma_error = sigma_error, interval = interval)
+
+  opt <- uniroot(
+    function(b0, ...) {
+      calc_power(c(b0, b1), ...)[1] - power
+    },
+    scal = scal, prog_vals = prog_vals, trt = trt, pred_vals = pred_vals, alpha = alpha,
+    type = type, sign_better = sign_better, sigma_error = sigma_error,
+    lambda0 = lambda0, cens_time = cens_time, t_mile = t_mile, interval = interval
+  )
   out <- opt$root
-  outpow <- calc_power(c(out,b1), scal=scal, prog_vals=prog_vals, trt=trt, pred_vals=pred_vals,
-                       alpha = alpha, type = type,
-                       sign_better = sign_better, sigma_error = sigma_error)[1]
+  outpow <- calc_power(c(out, b1),
+    scal = scal, prog_vals = prog_vals, trt = trt, pred_vals = pred_vals,
+    alpha = alpha, type = type,
+    sign_better = sign_better, sigma_error = sigma_error,
+    lambda0 = lambda0, cens_time = cens_time, t_mile = t_mile
+  )[1]
   attr(out, "power_results") <- outpow
-  if(abs(outpow-power) > 0.01)
+  if (abs(outpow - power) > 0.01) {
     message("optimization failed, try different starting value or optim_method")
+  }
   out
 }
-