+++ NEWS.md
@@ -1,3 +1,13 @@
+# benchtm 0.3.0 (Jan.8th 2024)
+
+* added time-to-event endpoints and corresponding scenarios
+
+##################################################
+# benchtm 0.2.0 (Nov. 4th, 2021)
+
+* generate new scenario cases where the overall power and power for detecting treatment effect heterogeneity are controlled at 0.50 and 0.80
+
+##################################################
 # benchtm 0.1.0
 
 * benchtm package on GitLab